requirements.txt
alembic==1.12.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Django==3.2.23
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.103.2
Flask==2.2.5
Flask-RESTful==0.3.10
greenlet==3.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.0.3
pydantic_core==2.14.5
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.2.3
zipp==3.15.0
